add ref to RFC2119 and href checker
Any good spec needs to use RFC2119 keywords, so this starts us down that path
by added a ref to it. Doesn't introduce any uses of the keywords yet.

Added an href checker tool so that we can make sure we don't have any
bad hrefs in our docs.

Also added a travis hook so we run the verification script as our build.

Signed-off-by: Doug Davis <dug@us.ibm.com>