Fixes #27
This attribute indicates the severity level of logging. The intent
of this attribute is to be considered as a mechanism to tell the receiving
infrastructure what the logging level of the infrastructure should be.

The log level can be one of debug, info, warning, error, fatal.
The set of log level is taken from syslog protocol.
Following is a description of each log level.

        fatal: system is unstable
        error: error conditions
        warning: warning conditions
        info : informational messages
	debug: debug-level messages

Signed-off-by: Malhar Vora <mlvora.2010@gmail.com>