Clarify the meaning of "source"
Related to #58, this is a follow-up to last week's meeting to clarify the meaning of source (fka resource).

The change is minimal and builds on the vocabulary of "occurrence" to be clearer that this is the first piece of software where the event originates, not the latest software to forward the event in a federated system. I've also added real examples from Google Firestore in case our interpretation helps others.

I'll add another PR to suggest an additional source-service.

Signed-off-by: Thomas Bouldin (inlined@google.com)