Proposes a new "service" sub-field of "source".
At Google we have found it useful to separate the service name
from the rest of the resource string:

* When registering an event trigger, the service name is a key into an address where we register our hooks. This service name can be a key into a private API hosted on another server address or on another protocol (e.g. gRPC)
* The service name can tersely be the only difference in trigger or event objects between:
  * Production, integration, and development tiers of a service.
  * Common APIs hosted across multiple services by a cloud provider (e.g. IAM access control).
  * Enterprise versions of software such as IBM OpenWhisk, GitHub, or HipChat.
  * Open-source services running in local infrastructure.

Edited the Google samples of source-id with the addition of source-service. I think there's multiple ways AWS could be interpreted and I leave it to a better expert to correct me.

Signed-off-by: Thomas Bouldin <inlined@google.com>