Introduce source-labels to help IoT cases.
Labels are a common concept in many existing products.
They are set on the source and are orthogonal to any
requests that are made against the resource. In the
events use cases, they can be particularly useful
for non-heirarchical filters or correlation of events.

It is reasoanble that the source itself is immutable
(E.g. a device UUID) but that an Action would respond
to a device's events based on where it is deployed.
(E.g. a particular user's house).

Signed-off-by: Thomas Bouldin <inlined@google.com>