Add practical constraints exercise to 0.3 roadmap.
In the 2018-06-15 f2f, we agreed to holistically revisit the spec
to look for issues that may arise with our spec in practice.

As a first swag, I placed this in the 0.3 milestone since it
has security-related issues. I could also see putting this in
the 0.4 milestone since it finalizes serialization and protocl mappings.

Signed-off-by: Thomas Bouldin <inlined@google.com>