Add some dates to our complete roadmap items
While it might be clear based on our version number where we are
in our plan/roadmap - I thought it might be good to put dates next
to each milestone.

For v0.1 it was easy, I just picked the date the release was cut.
For "Setup" it was a bit of a guess since its a unclear when we actually
complete ALL of those items. I picked Feb 26th since that was the date
we merged Dan's PR to add a LICENSE doc. It seemed as good of a date as
any - not sure it matters much.

Signed-off-by: Doug Davis <dug@us.ibm.com>