Add protobuf format (#295)
* Add protobuf format

This PR is meant for the spec as it is currently written, where the
extensions is an explicit bag. The focus of this PR is on the protobuf
format itself, not on the extensions proposal.

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Avoid redundant BytesWrapper by using bytes directly.

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Fix markdown links

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Reformat to be 80 columns.

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Only allow JSON data payloads

Efficient representation of binary bytes or proto payloads is not a
must have. The "data" field can represent bytes or serialized protos,
just in a less efficient matter.

I removed usages of these features to make the standard JSON look the
same as the CE JSON.

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Clarify attributing casing styles

protobuf lower_snake_case attrs are turned into lowerCamelCase in the
proto standard JSON form. But it is legal to have top level attribute
fields that are not lowerCamelCase.

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Update protobuf spec to use new design that favors forwards
compatibility over type safety.

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Fix URL to pass linter

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Fix whitespace

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Address PR comments, add example.

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Turn RFC references into links.

Signed-off-by: Spencer Fang <spencerfang@google.com>

* PR comments

- s/Attributes Atttributes/Attributes/
- s/required/mandated/
- Add protobuf-format.md to Makefile
- add cloudevent.proto

Signed-off-by: Spencer Fang <spencerfang@google.com>

* remove HTTP modes section

Signed-off-by: Spencer Fang <spencerfang@google.com>

* Reword confusing use of speficiation / below

Signed-off-by: Spencer Fang <spencerfang@google.com>