Add a version/timestamp attribute
Per Scott's suggestion, this adds a new attribute that can be used to
know when an entry has been updated. It can be used as an e-tag type of
thing or to know when, during an import, the entity being imported is
older than what the discovery endpoint already has.

I wasn't sure if a counter or a timestamp would be better, so I added
both so people can see how both look.

I'm leaning more towards a timestamp since it give a bit more information.

Signed-off-by: Doug Davis <dug@us.ibm.com>