Subscription OAS use of required and anyOf (#761)
* fix field id /subscriptions/{id} operation

path fields must have required=true according to OAS 3.0.3 and can not be omitted

Signed-off-by: Manuel Stein <manuel.stein@nokia-bell-labs.com>

* fix: request body is required in create and update

according to subscription spec (markdown), the content body of the
POST action to create a subscription is required. Also fixing required
content body in the PUT action that updates a subscription.

Signed-off-by: Manuel Stein <manuel.stein@nokia-bell-labs.com>

* fix protocolsettings parameter

Currently uses anyOf but in a parameter holding a single object, which
can hold only oneOf the provided schema objects. It's correct that only
a single protocol setting can be chosen per subscription.

Signed-off-by: Manuel Stein <manuel.stein@nokia-bell-labs.com>

* correcting protocolsettings

reverting oneOf to anyOf to allow ambiguously classifiable protocolsettings maps to be valid content

Signed-off-by: Manuel Stein <manuel.stein@nokia-bell-labs.com>