Initial attempt at batch clarification.
This treats batch-mode as a full-on third content mode (as it really
is). It attempts to clarify support in each transport and event
format, as well as removing duplication (e.g. in JSON format) of
what batch mode is about, and the odd "JSON batch format is a
separate event format" aspect.

Some parts may have been missed, and I'm thoroughly expecting some
push back :)

Signed-off-by: Jon Skeet <jonskeet@google.com>