Initial attempt at batch clarification (#1154)
* Initial attempt at batch clarification.

This treats batch-mode as a full-on third content mode (as it really
is). It attempts to clarify support in each transport and event
format, as well as removing duplication (e.g. in JSON format) of
what batch mode is about, and the odd "JSON batch format is a
separate event format" aspect.

Some parts may have been missed, and I'm thoroughly expecting some
push back :)

Signed-off-by: Jon Skeet <jonskeet@google.com>

* Address review feedback

(This includes removing some italics around content modes - we should probably do more of that, to be honest.)

Signed-off-by: Jon Skeet <jonskeet@google.com>

---------

Signed-off-by: Jon Skeet <jonskeet@google.com>