Clarify CDATA handling in XML format
(We should definitely have a big conformance suite for this format at some
point...)

The repetition of the language around XML element data is
deliberate, but slightly clunky. Other suggestions welcome.

Also happy to add an example with both CDATA and a text node if that
would be useful, but it feels like it's probably better in a
conformance test.

Signed-off-by: Jon Skeet <jonskeet@google.com>