Add unknown value for authtype enum of Auth Context ext (#1246)
* Change constraint to OPTIONAL for authtype

Issue created for this change: https://github.com/cloudevents/spec/issues/1245

For the recently added extension Auth Context /extensions/authcontext.md the attribute of authtype is the only one marked as REQUIRED in the constraints.

The key principle here is that authtype classification results should be predictable and should not change. However, for some cases the type is preferred to be unknown when we can't determine reliable. The main concern we have is how authtype classifications might change when we are able to classify the request in the future.
As a result if we change from unknown to app_user it is a "breaking" change for API consumers. Since their code may build business logic based on authtype results.

This is the reason we are suggesting making authtype OPTIONAL to avoid having customers build business logic around it.
If we have to return an authtype in a situation where we don't know for certain, I would prefer to add the enum value unknown.
We think is better to avoid this altogether until we can consistently and predictably classify "authtype" to avoid future "breaking change' scenario.

Signed-off-by: Josué Fabricio Urbina González <jfurbina@google.com>

* Add unknown value for authtype

From the discussion of the previous snapshot of github.com/cloudevents/spec/pull/1246

It is preferred to add another value to the authtype enum of `unknown`. This is to avoid setting the attribute as OPTIONAL since all extensions have at least one REQUIRED attribute.

Signed-off-by: Josué Fabricio Urbina González <jfurbina@google.com>

---------

Signed-off-by: Josué Fabricio Urbina González <jfurbina@google.com>